+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## 0.2.5
 
 * Allow proper form inheritance. When having `HitForm < SongForm < Reform::Form` the `HitForm` class will contain `SongForm`'s properties in addition to its own fields.
+* `::model` is now inherited properly.
 
 ## 0.2.4